- Removed unnecessary logger redundancy - Improved test coverage
