+++ pom.xml
@@ -3,7 +3,7 @@
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>es.upv.i3m.grycap</groupId>
 	<artifactId>im-java-api</artifactId>
-	<version>0.1.0</version>
+	<version>0.1.1</version>
 	<name>IM Java API</name>
 	<description>Java client for the REST API of the IM</description>
