+++ src/test/resources/log4j.properties
@@ -2,8 +2,6 @@
 log4j.rootLogger=ERROR
 
 # Application
-log4j.logger.es.upv.i3m.grycap.logger.GenericLogger=ERROR, CONSOLE
-log4j.additivity.es.upv.i3m.grycap.logger.GenericLogger=false
 log4j.logger.es.upv.i3m.grycap.logger.ImJavaApiLogger=DEBUG, CONSOLE
 log4j.additivity.es.upv.i3m.grycap.logger.ImJavaApiLogger=false
